+++ src/main/java/webserver/httpcontroller/HomeController.java
@@ -1,11 +1,11 @@
-package webserver.httphandler;
+package webserver.httpcontroller;
 
 import webserver.HttpRequest;
 import webserver.HttpResponse;
 
 import java.io.IOException;
 
-public class HomeHandler extends HttpHandler {
+public class HomeController extends HttpController {
     @Override
     public void service(HttpRequest request, HttpResponse response) throws IOException {
         response.setBody("Hello, World".getBytes(), "txt");
